refactor: move manifest specifics to the manifest.rs file under Manifest struct
