+++ src/environment.rs
@@ -92,7 +92,7 @@ pub async fn get_up_to_date_prefix(
     // Make sure the project supports the current platform
     let platform = Platform::current();
     if !project.platforms().contains(&platform) {
-        let span = project.manifest.project.platforms.span();
+        let span = project.manifest.manifest.project.platforms.span();
         return Err(miette::miette!(
             help = format!(
                 "The project needs to be configured to support your platform ({platform})."
@@ -103,7 +103,7 @@ pub async fn get_up_to_date_prefix(
             )],
             "the project is not configured for your current platform"
         )
-        .with_source_code(project.source()));
+        .with_source_code(project.manifest_named_source()));
     }
 
     // Make sure the system requirements are met
